requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
altair==5.5.0
annotated-types==0.7.0
anthropic==0.46.0
anyio==4.9.0
attrs==25.3.0
beautifulsoup4==4.13.4
blinker==1.9.0
cachetools==5.5.2
certifi==2025.4.26
cfgv==3.4.0
charset-normalizer==3.4.1
click==8.1.8
cmd2==2.5.11
contourpy==1.3.2
cycler==0.12.1
dataclasses-json==0.6.7
distlib==0.3.9
distro==1.9.0
entrypoints==0.4
faker==37.1.0
favicon==0.7.0
filelock==3.18.0
filetype==1.2.0
fonttools==4.57.0
frozenlist==1.6.0
fsspec==2025.3.2
ftfy==6.3.1
gitdb==4.0.12
gitpython==3.1.44
google-auth==2.39.0
google-genai==1.12.1
greenlet==3.2.1
h11==0.16.0
htbuilder==0.9.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.30.2
identify==2.6.10
idna==3.10
jinja2==3.1.6
jiter==0.9.0
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2025.4.1
kiwisolver==1.4.8
langchain==0.3.24
langchain-community==0.3.22
langchain-core==0.3.56
langchain-openai==0.3.14
langchain-text-splitters==0.3.8
langgraph==0.3.34
langgraph-checkpoint==2.0.24
langgraph-prebuilt==0.1.8
langgraph-sdk==0.1.63
langsmith==0.3.37
lxml==5.4.0
markdown==3.8
markdown-it-py==3.0.0
markdown2==2.5.3
markdownify==0.13.1
markdownlit==0.0.7
marker==2.1.3
marker-pdf==1.6.2
markupsafe==3.0.2
marshmallow==3.26.1
matplotlib==3.10.1
mdurl==0.1.2
mpmath==1.3.0
multidict==6.4.3
mypy-extensions==1.1.0
narwhals==1.36.0
networkx==3.4.2
nodeenv==1.9.1
numpy==2.2.5
nvidia-cublas-cu12==12.6.4.1
nvidia-cuda-cupti-cu12==12.6.80
nvidia-cuda-nvrtc-cu12==12.6.77
nvidia-cuda-runtime-cu12==12.6.77
nvidia-cudnn-cu12==9.5.1.17
nvidia-cufft-cu12==11.3.0.4
nvidia-cufile-cu12==1.11.1.6
nvidia-curand-cu12==10.3.7.77
nvidia-cusolver-cu12==11.7.1.2
nvidia-cusparse-cu12==12.5.4.2
nvidia-cusparselt-cu12==0.6.3
nvidia-nccl-cu12==2.26.2
nvidia-nvjitlink-cu12==12.6.85
nvidia-nvtx-cu12==12.6.77
openai==1.76.0
opencv-python-headless==4.11.0.86
orjson==3.10.16
ormsgpack==1.9.1
packaging==24.2
pandas==2.2.3
pdftext==0.6.2
pillow==10.4.0
platformdirs==4.3.7
plotly==6.0.1
pre-commit==4.2.0
prometheus-client==0.21.1
propcache==0.3.1
protobuf==5.29.4
pyarrow==19.0.1
pyasn1==0.6.1
pyasn1-modules==0.4.2
pydantic==2.11.3
pydantic-core==2.33.1
pydantic-settings==2.9.1
pydeck==0.9.1
pygments==2.19.1
pymdown-extensions==10.14.3
pyparsing==3.2.3
pypdfium2==4.30.0
pyperclip==1.9.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
pyyaml==6.0.2
rapidfuzz==3.13.0
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
rich==14.0.0
rpds-py==0.24.0
rsa==4.9.1
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.15.2
setuptools==79.0.1
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.7
sqlalchemy==2.0.40
st-annotated-text==4.0.2
st-theme==1.2.3
streamlit==1.44.1
streamlit-avatar==0.1.3
streamlit-camera-input-live==0.2.0
streamlit-card==1.0.2
streamlit-embedcode==0.1.2
streamlit-extras==0.6.0
streamlit-faker==0.0.3
streamlit-image-coordinates==0.1.9
streamlit-keyup==0.3.0
streamlit-toggle-switch==1.0.2
streamlit-vertical-slider==2.5.5
surya-ocr==0.13.1
sympy==1.13.3
tenacity==9.1.2
threadpoolctl==3.6.0
tiktoken==0.9.0
tokenizers==0.21.1
toml==0.10.2
torch==2.7.0
tornado==6.4.2
tqdm==4.67.1
transformers==4.51.3
triton==3.3.0
typing-extensions==4.13.2
typing-inspect==0.9.0
typing-inspection==0.4.0
tzdata==2025.2
urllib3==2.4.0
validators==0.34.0
virtualenv==20.30.0
watchdog==6.0.0
wcwidth==0.2.13
websockets==15.0.1
xxhash==3.5.0
yarl==1.20.0
zstandard==0.23.0
